+++ decomp/SDK/include/dolphin/mcc.h
@@ -48,7 +48,7 @@ enum MCC_SYSEVENT
 
 typedef void (*MCC_CBEvent)(enum MCC_CHANNEL channel, u32, u32);
 typedef void (*MCC_CBSysEvent)(enum MCC_SYSEVENT sysEvent);
-typedef int  (*MCC_CBEnumDevices)(long);
+typedef int  (*MCC_CBEnumDevices)(s32);
 
 typedef struct
 {
@@ -134,7 +134,7 @@ struct FIO_Stat
 struct FIO_Finddata
 {
     /* 0x00 */ struct FIO_Stat stat;
-    /* 0x24 */ char            filename[256];
+    /* 0x24 */ s8            filename[256];
 };
 
 // fio.c
@@ -146,7 +146,7 @@ int  FIOFopen(const char* filename, u32 mode);
 int  FIOFclose(int handle);
 u32  FIOFread(int handle, void* data, u32 size);
 u32  FIOFwrite(int handle, void* data, u32 size);
-u32  FIOFseek(int handle, long offset, u32 mode);
+u32  FIOFseek(int handle, s32 offset, u32 mode);
 int  FIOFprintf(int handle, const char* format, ...);
 int  FIOFflush(int handle);
 int  FIOFstat(int handle, struct FIO_Stat* stat);
@@ -177,8 +177,8 @@ int  MCCOpen(enum MCC_CHANNEL chID, u8 blockSize, MCC_CBEvent callbackEvent);
 int  MCCClose(enum MCC_CHANNEL chID);
 int  MCCLock(enum MCC_CHANNEL chID);
 int  MCCUnlock(enum MCC_CHANNEL chID);
-int  MCCRead(enum MCC_CHANNEL chID, u32 offset, void* data, long size, enum MCC_SYNC_STATE async);
-int  MCCWrite(enum MCC_CHANNEL chID, u32 offset, void* data, long size, enum MCC_SYNC_STATE async);
+int  MCCRead(enum MCC_CHANNEL chID, u32 offset, void* data, s32 size, enum MCC_SYNC_STATE async);
+int  MCCWrite(enum MCC_CHANNEL chID, u32 offset, void* data, s32 size, enum MCC_SYNC_STATE async);
 int  MCCCheckAsyncDone();
 
 #endif // _DOLPHIN_MCC_H_
