+++ decomp/SDK/include/dolphin/os/OSAlarm.h
@@ -22,7 +22,7 @@ BOOL OSCheckAlarmQueue(void);
 void OSInitAlarm(void);
 void OSCreateAlarm(OSAlarm* alarm);
 void OSSetAlarm(OSAlarm* alarm, OSTime tick, OSAlarmHandler handler);
-void OSSetAbsAlarm(struct OSAlarm* alarm, long long time, void (*handler)(struct OSAlarm*, struct OSContext*));
+void OSSetAbsAlarm(struct OSAlarm* alarm, s64 time, void (*handler)(struct OSAlarm*, struct OSContext*));
 void OSSetPeriodicAlarm(OSAlarm* alarm, OSTime start, OSTime period, OSAlarmHandler handler);
 void OSCancelAlarm(OSAlarm* alarm);
