+++ decomp/SDK/src/dolphin/gx/GXMisc.c
@@ -499,8 +499,8 @@ GXDecompressZ16(u32 z16, GXZFmt16 zfmt)
 {
     u32  z24;
     u32  cb1;
-    long exp;
-    long shift;
+    s32 exp;
+    s32 shift;
 
     cb1;
     cb1;
